RPC helper: payload receiver's size_t check updated

+++ include/rpc_utils.h
@@ -13,9 +13,9 @@
 
 uint64_t hash(unsigned char* str);
 int rpc_receive_uint(int socket, uint64_t* ret);
-int rpc_receive_int(int socket, int64_t* ret);
+int rpc_receive_int(int socket, int* ret);
 int rpc_send_uint(int socket, uint64_t val);
-int rpc_send_int(int socket, int64_t val);
+int rpc_send_int(int socket, int val);
 
 int rpc_send_payload(int socket, rpc_data* payload);
 rpc_data* rpc_receive_payload(int socket);
